feat: update method getProductCount

+++ product-service/src/main/java/vantruong/productservice/service/ProductService.java
@@ -31,5 +31,5 @@ public interface ProductService {
 
   List<Product> getAll();
 
-  int getProductCount();
+  Long getProductCount();
 }
